Code refactor with useAppSelectDispatch.
